refactor: centralize custom element declarations
