+++ src/asset.ts
@@ -314,11 +314,4 @@ class AssetElement extends AsyncElement {
 
 customElements.define('pc-asset', AssetElement);
 
-declare global {
-    // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-    interface HTMLElementTagNameMap {
-        'pc-asset': AssetElement;
-    }
-}
-
 export { AssetElement };
