+++ src/components/sound-slot.ts
@@ -311,11 +311,4 @@ class SoundSlotElement extends AsyncElement {
 
 customElements.define('pc-sound', SoundSlotElement);
 
-declare global {
-    // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-    interface HTMLElementTagNameMap {
-        'pc-sound': SoundSlotElement;
-    }
-}
-
 export { SoundSlotElement };
