+++ src/module.ts
@@ -50,11 +50,4 @@ class ModuleElement extends HTMLElement {
 
 customElements.define('pc-module', ModuleElement);
 
-declare global {
-    // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
-    interface HTMLElementTagNameMap {
-        'pc-module': ModuleElement;
-    }
-}
-
 export { ModuleElement };
